Week-35/day-5/requirements.txt
sqlalchemy
flask
flask-wtf
wtforms
python-dotenv
email-validator
flask-sqlalchemy
psycopg2-binary
alembic
flask-migrate
